+++ apps/meteor/client/sidebar/header/CreateDirectMessage.tsx
@@ -1,7 +1,7 @@
 import type { IUser } from '@rocket.chat/core-typings';
 import { Box, Modal, Button, FieldGroup, Field, FieldRow, FieldLabel, FieldError } from '@rocket.chat/fuselage';
 import { useUniqueId } from '@rocket.chat/fuselage-hooks';
-import { useTranslation, useEndpoint, useToastMessageDispatch } from '@rocket.chat/ui-contexts';
+import { useTranslation, useEndpoint, useToastMessageDispatch, useSetting } from '@rocket.chat/ui-contexts';
 import { useMutation } from '@tanstack/react-query';
 import React, { memo } from 'react';
 import { useForm, Controller } from 'react-hook-form';
@@ -11,6 +11,7 @@ import { goToRoomById } from '../../lib/utils/goToRoomById';
 
 const CreateDirectMessage = ({ onClose }: { onClose: () => void }) => {
 	const t = useTranslation();
+	const directMaxUsers = useSetting<number>('DirectMesssage_maxUsers') || 1;
 	const membersFieldId = useUniqueId();
 	const dispatchToastMessage = useToastMessageDispatch();
 
@@ -55,7 +56,13 @@ const CreateDirectMessage = ({ onClose }: { onClose: () => void }) => {
 						<FieldRow>
 							<Controller
 								name='users'
-								rules={{ required: t('error-the-field-is-required', { field: t('Members') }) }}
+								rules={{
+									required: t('error-the-field-is-required', { field: t('Members') }),
+									validate: (users) =>
+										users.length + 1 > directMaxUsers
+											? t('error-direct-message-max-user-exceeded', { maxUsers: directMaxUsers })
+											: undefined,
+								}}
 								control={control}
 								render={({ field: { name, onChange, value, onBlur } }) => (
 									<UserAutoCompleteMultipleFederated