+++ apps/meteor/server/startup/migrations/xrun.ts
@@ -6,10 +6,11 @@ const { MIGRATION_VERSION = 'latest' } = process.env;
 
 const [version, ...subcommands] = MIGRATION_VERSION.split(',');
 
-await migrateDatabase(version === 'latest' ? version : parseInt(version), subcommands);
-
-// perform operations when the server is starting with a different version
-await onServerVersionChange(async () => {
-	await upsertPermissions();
-	await ensureCloudWorkspaceRegistered();
-});
+export const performMigrationProcedure = async (): Promise<void> => {
+	await migrateDatabase(version === 'latest' ? version : parseInt(version), subcommands);
+	// perform operations when the server is starting with a different version
+	await onServerVersionChange(async () => {
+		await upsertPermissions();
+		await ensureCloudWorkspaceRegistered();
+	});
+};