+++ apps/meteor/tests/e2e/page-objects/auth.ts
@@ -20,6 +20,14 @@ export class Registration {
 		return this.page.locator('role=button[name="Login"]');
 	}
 
+	get btnLoginWithSaml(): Locator {
+		return this.page.locator('role=button[name="SAML"]');
+	}
+
+	get btnLoginWithGoogle(): Locator {
+		return this.page.locator('role=button[name="Sign in with Google"]');
+	}
+
 	get goToRegister(): Locator {
 		return this.page.locator('role=link[name="Create an account"]');
 	}