+++ packages/core-services/src/events/Events.ts
@@ -18,7 +18,6 @@ import type {
 	ISocketConnection,
 	ISubscription,
 	IUser,
-	IUserStatus,
 	IInvite,
 	IWebdavAccount,
 	ICustomSound,
@@ -33,12 +32,26 @@ import type {
 	IBanner,
 	ILivechatVisitor,
 	LicenseLimitKind,
+	ICustomUserStatus,
 } from '@rocket.chat/core-typings';
 import type * as UiKit from '@rocket.chat/ui-kit';
 
 import type { AutoUpdateRecord } from '../types/IMeteor';
 
 type ClientAction = 'inserted' | 'updated' | 'removed' | 'changed';
+
+type LoginServiceConfigurationEvent = {
+	id: string;
+} & (
+	| {
+			clientAction: 'removed';
+			data?: never;
+	  }
+	| {
+			clientAction: Omit<ClientAction, 'removed'>;
+			data: Partial<ILoginServiceConfiguration>;
+	  }
+);
 
 export type EventSignatures = {
 	'room.video-conference': (params: { rid: string; callId: string }) => void;
@@ -115,7 +128,7 @@ export type EventSignatures = {
 					replaceByUser: { _id: IUser['_id']; username: IUser['username']; alias: string };
 			  },
 	): void;
-	'user.deleteCustomStatus'(userStatus: IUserStatus): void;
+	'user.deleteCustomStatus'(userStatus: Omit<ICustomUserStatus, '_updatedAt'>): void;
 	'user.nameChanged'(user: Pick<IUser, '_id' | 'name' | 'username'>): void;
 	'user.realNameChanged'(user: Partial<IUser>): void;
 	'user.roleUpdate'(update: {
@@ -124,7 +137,7 @@ export type EventSignatures = {
 		u?: { _id: IUser['_id']; username: IUser['username']; name?: IUser['name'] };
 		scope?: string;
 	}): void;
-	'user.updateCustomStatus'(userStatus: IUserStatus): void;
+	'user.updateCustomStatus'(userStatus: Omit<ICustomUserStatus, '_updatedAt'>): void;
 	'user.typing'(data: { user: Partial<IUser>; isTyping: boolean; roomId: string }): void;
 	'user.video-conference'(data: {
 		userId: IUser['_id'];
@@ -139,7 +152,7 @@ export type EventSignatures = {
 		user: Pick<IUser, '_id' | 'username' | 'status' | 'statusText' | 'name' | 'roles'>;
 		previousStatus: UserStatus | undefined;
 	}): void;
-	'watch.messages'(data: { clientAction: ClientAction; message: IMessage }): void;
+	'watch.messages'(data: { message: IMessage }): void;
 	'watch.roles'(
 		data:
 			| { clientAction: Exclude<ClientAction, 'removed'>; role: IRole }
@@ -235,7 +248,7 @@ export type EventSignatures = {
 			  }
 		),
 	): void;
-	'watch.loginServiceConfiguration'(data: { clientAction: ClientAction; data: Partial<ILoginServiceConfiguration>; id: string }): void;
+	'watch.loginServiceConfiguration'(data: LoginServiceConfigurationEvent): void;
 	'watch.instanceStatus'(data: {
 		clientAction: ClientAction;
 		data?: undefined | Partial<IInstanceStatus>;
@@ -287,5 +300,4 @@ export type EventSignatures = {
 	'command.updated'(command: string): void;
 	'command.removed'(command: string): void;
 	'actions.changed'(): void;
-	'message.sent'(message: IMessage): void;
 };